# Conflicts:
#	api-test/post-test.http
#	src/main/java/com/sparta/restplaceforj/entity/Post.java
#	src/main/java/com/sparta/restplaceforj/exception/ErrorEnum.java
#	src/main/java/com/sparta/restplaceforj/service/PostService.java

+++ src/main/java/com/sparta/restplaceforj/service/PostService.java
@@ -8,6 +8,7 @@ import com.sparta.restplaceforj.dto.PostResponseDto;
 import com.sparta.restplaceforj.entity.Image;
 import com.sparta.restplaceforj.entity.Post;
 import com.sparta.restplaceforj.entity.ThemeEnum;
+import com.sparta.restplaceforj.entity.User;
 import com.sparta.restplaceforj.exception.CommonException;
 import com.sparta.restplaceforj.exception.ErrorEnum;
 import com.sparta.restplaceforj.repository.ImageRepository;
@@ -45,10 +46,14 @@ public class PostService {
   private final S3Service s3Service;
 
   /**
-   * 글 새성.
+   * 글 생성
+   *
+   * @param postRequestDto : title, content, address, theme, placeName
+   * @param user           : 토큰 유저 정보
+   * @return PostResponseDto : title, content, address, likesCount, viewCount, themeEnum
    */
   @Transactional
-  public PostResponseDto createPost(PostRequestDto postRequestDto) {
+  public PostResponseDto createPost(PostRequestDto postRequestDto, User user) {
 
     try {
       ThemeEnum.valueOf(postRequestDto.getTheme());
@@ -59,6 +64,7 @@ public class PostService {
     Post post = postRepository.save(
         Post.builder()
             .requestDto(postRequestDto)
+            .user(user)
             .build());
 
     if (!postRequestDto.getImageIdList().isEmpty()) {
@@ -75,8 +81,19 @@ public class PostService {
         .build();
   }
 
+  /**
+   * 삭제
+   *
+   * @param postId pathVariable
+   * @param user   : 토큰 유저 정보
+   */
   @Transactional
-  public void deletePost(long postId) {
+  public void deletePost(long postId, User user) {
+    Post post = postRepository.findByIdOrThrow(postId);
+    if (post.getUser().getId() != user.getId()) {
+      throw new CommonException(ErrorEnum.POST_MISMATCH);
+    }
+
     postRepository.deleteById(postId);
   }
 
@@ -95,7 +112,14 @@ public class PostService {
   }
 
   /**
-   * 글 아이디와 제목만 조회.
+   * 글을 조회하면 제목과 아이디만 반환  .
+   *
+   * @param page      현재 페이지
+   * @param size      페이지 크기
+   * @param placeName 장소명
+   * @param q         검색질문
+   * @param sortBy    정렬 기준
+   * @return PageResponseDto : placeNameList, size, page, totalPages, totalElements
    */
   public PageResponseDto<PostIdTitleDto> getPostTitleList(
       int page, int size, String placeName, String sortBy, String q) {
@@ -118,11 +142,20 @@ public class PostService {
   }
 
   /**
-   * 글 수정.
+   * 글 수정
+   *
+   * @param postId         수정 글 아이디
+   * @param postRequestDto title, content, address, theme, placeName
+   * @param user           : 토큰 유저 정보
+   * @return PostResponseDto id, userId, title, content, address, likesCount, viewsCount, themeEnum
    */
   @Transactional
-  public PostResponseDto updatePost(long postId, PostRequestDto postRequestDto) {
+  public PostResponseDto updatePost(long postId, PostRequestDto postRequestDto, User user) {
     Post post = postRepository.findByIdOrThrow(postId);
+    if (post.getUser().getId() != user.getId()) {
+      throw new CommonException(ErrorEnum.POST_MISMATCH);
+    }
+
     post.update(postRequestDto);
     return PostResponseDto.builder()
         .post(post)
@@ -130,10 +163,15 @@ public class PostService {
   }
 
   /**
-   * 글 단권 조회.
+   * 단권글 조회.
+   *
+   * @param postId 조회 글 아이디
+   * @return PostResponseDto :id, userId, title, content, address, likesCount, viewsCount, themeEnum
    */
+  @Transactional
   public PostResponseDto getPost(long postId) {
     Post post = postRepository.findByIdOrThrow(postId);
+    post.addViewToPost();
 
     return PostResponseDto.builder()
         .post(post)
